MedBot-RAG-main/requirements.txt
llama_index==0.10.56
python-dotenv==1.0.1
streamlit==1.31.1
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
